requirements.txt
openai==1.2
requests
beautifulsoup4
fastapi
uvicorn
python-dotenv
pypdf2